feat: make `adapter` configurable in the `home` view (#750)
EX-7073
