+++ packages/x-components/src/views/base-config.ts
@@ -1,7 +1,6 @@
-import { platformAdapter } from '@empathyco/x-adapter-platform';
 import { SnippetConfig } from '../x-installer/api/api.types';
 import { InstallXOptions } from '../x-installer/x-installer/types';
-import { e2eAdapter } from '../adapter/e2e-adapter';
+import { adapter } from './adapter';
 
 export const baseSnippetConfig: SnippetConfig = {
   instance: 'empathy',
@@ -9,8 +8,6 @@ export const baseSnippetConfig: SnippetConfig = {
   env: 'staging',
   scope: 'x-components-development'
 };
-
-const adapter = 'Cypress' in window ? e2eAdapter : platformAdapter;
 
 const xModulesURLConfig = JSON.parse(new URL(location.href).searchParams.get('xModules') ?? '{}');
